requirements.txt
pandas>=1.3.0
pyyaml>=6.0
jinja2>=3.0
click>=8.0
rich>=12.0
colorama>=0.4.0
pyspark>=3.0.0
sqlalchemy>=1.4.0
pytest>=7.0
pytest-cov>=4.0
